move SeedChooserScreen out
