+++ pvzclass/Classes/SeedChoserScreen.cpp
@@ -1,4 +1,4 @@
-#include "../PVZ.h"
+#include "SeedChooserScreen.hpp"
 
 PVZ::Board PVZ::SeedChooserScreen::GetBoard()
 {
@@ -8,4 +8,10 @@ PVZ::Board PVZ::SeedChooserScreen::GetBoard()
 PVZ::SeedChooserScreen::ChosenSeed PVZ::SeedChooserScreen::GetChosenSeed(int num)
 {
 	return(PVZ::SeedChooserScreen::ChosenSeed(BaseAddress + 0xA4 + num * 0x3C));
+}
+
+PVZ::SeedChooserScreen PVZ::GetSeedChooserScreen()
+{
+	int address = Memory::ReadPointer(0x6A9EC0, 0x774);
+	return(address == 0 ? INVALID_BASEADDRESS : SeedChooserScreen(address));
 }
