+++ pvzclass/pvzclass.vcxproj
@@ -245,6 +245,7 @@
     <ClInclude Include="Classes\ChallengeScreen.hpp" />
     <ClInclude Include="Classes.hpp" />
     <ClInclude Include="Classes\ResourceManager.hpp" />
+    <ClInclude Include="Classes\SeedChooserScreen.hpp" />
     <ClInclude Include="Classes\TodParticleSystem.hpp" />
     <ClInclude Include="Draw.h" />
     <ClInclude Include="Enums\GameObjectType.hpp" />
